feat: chart updates
